bug: 추가등록자 리뷰 필수조건 수정(시그니처 메뉴 필수 아님)
