+++ breadgood_app/lib/modules/main/controller/main_map_controller.dart
@@ -9,8 +9,8 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:naver_map_plugin/naver_map_plugin.dart';
 import 'package:http/http.dart' as http;
-import 'package:breadgood_app/constant/api_path.dart' as PATH;
-import 'package:breadgood_app/utils/services/rest_api_service.dart' as REST_API;
+import 'package:breadgood_app/constant/api_path.dart' as api_path;
+import 'package:breadgood_app/utils/services/rest_api_service.dart' as rest_api;
 
 class MainMapController extends GetxController {
   Completer<NaverMapController> naverController = Completer();
@@ -71,8 +71,8 @@ class MainMapController extends GetxController {
   }
 
   Future<List> bakeryList(city,district) async {
-    final url = Uri.parse("${PATH.REST_API_URL}/bakery/search");
-    Map<String,String> headers = await REST_API.headers();
+    final url = Uri.parse("${api_path.restApiUrl}/bakery/search");
+    Map<String,String> headers = await rest_api.headers();
     // var body=json.encode({"city": city, "district": district, "bakeryCategory1": "1"});
     var body=json.encode({"city": city, "district": district, "bakeryCategory": null});
     if(city=="서울특별시"){
