+++ breadgood_app/lib/modules/main/screens/bread_store_deatil_webview.dart
@@ -1,8 +1,7 @@
 import 'package:breadgood_app/modules/register_bakery/screens/already_registered_bakery.dart';
 import 'package:breadgood_app/utils/ui/bottomNavigation.dart';
 import 'package:flutter/material.dart';
-// import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
-import 'package:breadgood_app/constant/api_path.dart' as PATH;
+import 'package:breadgood_app/constant/api_path.dart' as api_path;
 import 'package:get/get.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
@@ -10,7 +9,7 @@ final Set<JavascriptChannel> jsChannels = [
   JavascriptChannel(
       name: 'MoveToReview',
       onMessageReceived: (JavascriptMessage message) {
-        Get.toNamed("/register_review/register_reivew_page");
+        Get.toNamed("/register_review/register_reivew_page", arguments: Get.arguments['bakeryId']);
       }),
 ].toSet();
 
@@ -27,7 +26,7 @@ class _BreadStoreDetailWebViewState extends State<BreadStoreDetailWebView> {
       appBar: AlreadyRegisteredBakeryAppbar(),
       body: WebView(
         initialUrl:
-            "${PATH.URL}/pages/bakery/detail?userId=${Get.arguments['userId']}&bakeryId=${Get.arguments['bakeryId']}",
+            "${api_path.url}/pages/bakery/detail?userId=${Get.arguments['userId']}&bakeryId=${Get.arguments['bakeryId']}",
         javascriptMode: JavascriptMode.unrestricted,
         javascriptChannels: jsChannels,
       ),
