+++ breadgood_app/lib/modules/register_bakery/screens/already_registered_bakery.dart
@@ -1,6 +1,6 @@
 import 'package:breadgood_app/modules/main/screens/main_map.dart';
 import 'package:breadgood_app/modules/register_bakery/screens/search_bakery.dart';
-import 'package:breadgood_app/modules/register_review/register_review.dart';
+import 'package:breadgood_app/modules/register_review/screens/register_review.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:breadgood_app/utils/ui/main_app_bar.dart';
@@ -23,11 +23,10 @@ class _AlreadyRegisteredBakeryPageState
     extends State<AlreadyRegisteredBakeryPage> {
   final controller = Get.put(BakeryController());
 
-  // var args = Get.arguments;
   SearchData selectedBakery;
-  // = controller.selectedBakery;
-  String first_registerer = Get.arguments;
-  // String first_registerer = '동오언니';
+  String first_registerer = (Get.arguments)[0];
+  int bakeryId = (Get.arguments)[1];
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -131,7 +130,7 @@ class _AlreadyRegisteredBakeryPageState
                                         fontWeight: FontWeight.w600,
                                       )),
                                   onPressed: () {
-                                    Get.to(RegisterReviewPage());
+                                    Get.to(RegisterReviewPage(), arguments: bakeryId);
                                   },
                                 ),
                               ),
